+++ src/database/migrations/20211004110017-relation-user-currencies.js
@@ -29,11 +29,11 @@ module.exports = {
         onDelete: 'CASCADE'
       },
       currency_quantity: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(10,2),
         allowNull: false
       },
       currency_price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(10,2),
         allowNull: false
       },
       created_at: {
